Add fraction-size option to directive to round cents.
Parts:

- Add npm dev-dependencies and npm test command
- Replace $filter with currencyFilter directly
- Upgrade angular to 1.3 for currencyFilter fractionSize param.

I changed some tests related to -0 conversion.  I belive currencyFilter
changed its implementation in angular 1.3, and that these tests were to
document the behavior but weren't actually a feature people relied upon.

https://github.com/angular/angular.js/blob/master/CHANGELOG.md#130-superluminal-nudge-2014-10-13

I am not totally sure if the rounding should go in both directions.
E.g. if the user enters '1.99' it is currently storing on the model
as 1.99, but displayed as 2 when fraction-size is 0.